added jest as unit test framework
