+++ .eslintrc
@@ -5,7 +5,8 @@
   
     "env": {
       "node": true,
-      "es6": true
+      "es6": true,
+      "jest": true
     },
   
     "rules": {
