+++ .travis.yml
@@ -10,3 +10,4 @@ before_script:
 
 script:
   - npm run lint
+  - npm run test
