+++ CHANGES.txt
@@ -1,3 +1,7 @@
+## 2.0.0 (, 2019)
+
+* Added jest as unit test framework.
+
 ## 1.4.5 (October 8, 2018)
 
 * Updating SDK version to v10.4.0 which is the latest stable to the date. Adding tilde.
