+++ package.json
@@ -19,10 +19,14 @@
       "url": "https://github.com/mmelograno"
     }
   ],
+  "jest": {
+    "testPathIgnorePatterns": ["config"]
+  },
   "main": "server.js",
   "scripts": {
     "start": "node server.js",
-    "lint": "eslint ."
+    "lint": "eslint .",
+    "test": "NODE_ENV=test SPLITIO_API_KEY=localhost EXT_API_KEY=test jest"
   },
   "dependencies": {
     "@splitsoftware/splitio": "~10.8.2",
@@ -35,6 +39,12 @@
     "npm": "6.10.1"
   },
   "devDependencies": {
-    "eslint": "^6.1.0"
+    "@babel/core": "^7.5.5",
+    "@babel/preset-env": "^7.5.5",
+    "babel-jest": "^24.8.0",
+    "eslint": "^6.1.0",
+    "jest": "^24.8.0",
+    "superagent": "^5.1.0",
+    "supertest": "^4.0.2"
   }
 }
